+++ app/assets/javascripts/application.js
@@ -3,8 +3,4 @@
 //= require jquery
 //= require moment
 //= require moment/fr.js
-import 'mapbox-gl/dist/mapbox-gl.css';
 
-import { initMapbox } from '../plugins/init_mapbox';
-
-initMapbox();
