+++ app/javascript/packs/application.js
@@ -1,1 +1,6 @@
 import "bootstrap";
+import 'mapbox-gl/dist/mapbox-gl.css';
+
+import { initMapbox } from '../plugins/init_mapbox';
+
+initMapbox();
