+++ package.json
@@ -4,9 +4,10 @@
   "dependencies": {
     "@rails/webpacker": "3.5",
     "bootstrap": "3.4.0",
-    "jquery": "^3.3.1"
+    "jquery": "^3.3.1",
+    "mapbox-gl": "^0.53.0"
   },
   "devDependencies": {
-    "webpack-dev-server": ">=3.1.11"
+    "webpack-dev-server": "2.11.2"
   }
 }
